+++ dataiter/vector.py
@@ -303,6 +303,16 @@ class Vector(np.ndarray):
         self._check_dimensions()
         return self.size
 
+    def map(self, function, *args, **kwargs):
+        """
+        Apply `function` element-wise and return a new vector.
+
+        >>> import math
+        >>> vector = di.Vector(range(10))
+        >>> vector.map(math.pow, 2)
+        """
+        return self.__class__(function(x, *args, **kwargs) for x in self)
+
     @property
     def missing_dtype(self):
         """