+++ NEWS.md
@@ -2,6 +2,7 @@ PENDING: dataiter 0.18
 ======================
 
 * `Vector.__init__`: Accept generators/iterators
+* `Vector.map`: New method
 
 2021-02-27: dataiter 0.17
 =========================